Remove pageState Rename viewModel
